meeting-summarizer-app/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
openai-whisper
ollama
pydantic>=2.9.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
torch
torchaudio
sqlalchemy>=2.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
websockets>=11.0.3
asyncio-mqtt>=0.16.1
